We have reset memcpy()-like functions to use the prototype that libdwarf has long exported to callers (on request).  void  copy_word(void *, const void *,unsigned long)
The length in practice is small where our swap/no-swap
functions are involved.

There were some inconsistencies but freebsd 32bit
with clang caught them all, so it should be ok
everywhere.
	modified:   src/lib/libdwarf/dwarf_machoread.c
	modified:   src/lib/libdwarf/dwarf_memcpy_swap.c
	modified:   src/lib/libdwarf/dwarf_memcpy_swap.h
	modified:   src/lib/libdwarf/dwarf_object_detector.c
	modified:   src/lib/libdwarf/dwarf_peread.c

+++ src/lib/libdwarf/dwarf_machoread.c
@@ -1141,7 +1141,7 @@ static int
 fill_in_uni_arch_32(
     struct fat_arch * fa,
     struct Dwarf_Universal_Head_s *duhd,
-    void (*word_swap) (void *, const void *, size_t))
+    void (*word_swap) (void *, const void *, unsigned long))
 {
     Dwarf_Unsigned i = 0;
     struct Dwarf_Universal_Arch_s * dua = 0;
@@ -1162,7 +1162,7 @@ static int
 fill_in_uni_arch_64(
     struct fat_arch_64 * fa,
     struct Dwarf_Universal_Head_s *duhd,
-    void (*word_swap) (void *, const void *, size_t))
+    void (*word_swap) (void *, const void *, unsigned long))
 {
     Dwarf_Unsigned i = 0;
     struct Dwarf_Universal_Arch_s * dua = 0;
@@ -1194,7 +1194,7 @@ _dwarf_object_detector_universal_head_fd(
     struct Dwarf_Universal_Head_s *duhdp = 0;
     struct  fat_header fh;
     int     res = 0;
-    void (*word_swap) (void *, const void *, size_t);
+    void (*word_swap) (void *, const void *, unsigned long);
     int     locendian = 0;
     int     locoffsetsize = 0;
